ceph-disk: avoid remove '/var/lib/ceph/osd/ceph-xx' directory when deactivate with --once   when using deactivate with --once, '/var/lib/ceph/osd/ceph-xx' will be deleted by call umount,   but '--once' means to keep that directory for reactivate.

+++ src/ceph-disk/ceph_disk/main.py
@@ -1376,6 +1376,7 @@ def mount(
 
 def unmount(
     path,
+    dont_rm=False,
 ):
     """
     Unmount and removes the given mount point.
@@ -1399,7 +1400,8 @@ def unmount(
             else:
                 time.sleep(0.5 + retries * 1.0)
                 retries += 1
-
+    if dont_rm:
+        return
     os.rmdir(path)
 
 
@@ -3787,7 +3789,7 @@ def main_deactivate_locked(args):
         with open(os.path.join(mounted_path, 'deactive'), 'w'):
             path_set_context(os.path.join(mounted_path, 'deactive'))
 
-    unmount(mounted_path)
+    unmount(mounted_path, args.once)
     LOG.info("Umount `%s` successfully.", mounted_path)
 
     if dmcrypt:
